+++ frontend/app/keyboard/page.tsx
@@ -3,7 +3,7 @@
 import useKeyboardPage from "@/hooks/useKeyboardPage";
 
 const KeyboardPage = () => {
-  const { keyboard, keyMap } = useKeyboardPage();
+  const { keyboard, currentPressedKeys, allPressedKeys } = useKeyboardPage();
   return (
     <div className="flex flex-col gap-16">
       <h1 className="text-xl font-bold">Key Tester</h1>
@@ -16,7 +16,7 @@ const KeyboardPage = () => {
             >
               {row.map((key, index) => (
                 <kbd
-                  className={`kbd ${keyMap.get(key.value) && "bg-accent text-accent-content"}`}
+                  className={`kbd ${currentPressedKeys.has(key.value) ? "bg-accent/60" : allPressedKeys.has(key.value) && "bg-accent text-accent-content"}`}
                   style={{
                     minWidth: key.width * 48,
                     minHeight: 48,
@@ -29,7 +29,6 @@ const KeyboardPage = () => {
             </div>
           ))}
         </div>
-        
       </div>
     </div>
   );
