+++ frontend/types/keyboard.d.ts
@@ -1,0 +1,5 @@
+export interface Key {
+  value: string;
+  label: string;
+  width: number;
+}
